tests: basic E2E framework base
