+++ .circleci/config.yml
@@ -14,8 +14,6 @@ jobs:
           machine: true
       - shared/with_go_cache
       - shared/save_go_cache
-      - shared/setup_devenv:
-          provision_args: "--base"
       - run:
           name: Run E2E tests
           command: make e2e-override
@@ -115,7 +113,6 @@ workflows:
             - ghaccesstoken
             - npm-credentials
             - vault-dev
-            - aws-credentials
       ###EndBlock(circleWorkflowJobs)
       - release:
           context:
