+++ cmd/devenv/snapshot/snapshot.go
@@ -1,24 +1,41 @@
 package snapshot
 
 import (
+	"archive/tar"
+	"bytes"
 	"context"
+	"crypto/md5"
+	"encoding/base64"
 	"fmt"
+	"io"
 	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
 	"strings"
 	"time"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
 	dockerclient "github.com/docker/docker/client"
+	"github.com/getoutreach/devenv/cmd/devenv/destroy"
+	"github.com/getoutreach/devenv/cmd/devenv/provision"
+	devenvaws "github.com/getoutreach/devenv/pkg/aws"
 	"github.com/getoutreach/devenv/pkg/cmdutil"
+	"github.com/getoutreach/devenv/pkg/devenvutil"
 	"github.com/getoutreach/devenv/pkg/kube"
-	"github.com/getoutreach/devenv/pkg/worker"
+	"github.com/getoutreach/devenv/pkg/kubernetesruntime"
+	noncmdsnapshot "github.com/getoutreach/devenv/pkg/snapshot"
+	"github.com/getoutreach/devenv/pkg/snapshoter"
 	"github.com/getoutreach/gobox/pkg/box"
+	"github.com/minio/minio-go/v7"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
 	"gopkg.in/yaml.v2"
 
-	corev1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -28,10 +45,6 @@ import (
 	veleroclient "github.com/vmware-tanzu/velero/pkg/generated/clientset/versioned"
 	velerov1 "github.com/vmware-tanzu/velero/pkg/generated/informers/externalversions/velero/v1"
 	"github.com/vmware-tanzu/velero/pkg/util/boolptr"
-)
-
-const (
-	SnapshotNamespace = "velero"
 )
 
 //nolint:gochecknoglobals
@@ -56,10 +69,11 @@ type Options struct {
 	k   kubernetes.Interface
 	r   *rest.Config
 	d   dockerclient.APIClient
+	b   *box.Config
 	vc  veleroclient.Interface
 }
 
-func NewOptions(log logrus.FieldLogger) (*Options, error) {
+func NewOptions(log logrus.FieldLogger, b *box.Config) (*Options, error) {
 	k, conf, err := kube.GetKubeClientWithConfig()
 	if err != nil {
 		log.WithError(err).Warn("failed to create kubernetes client")
@@ -72,6 +86,7 @@ func NewOptions(log logrus.FieldLogger) (*Options, error) {
 
 	opts := &Options{
 		log: log,
+		b:   b,
 		d:   d,
 	}
 
@@ -91,16 +106,10 @@ func NewOptions(log logrus.FieldLogger) (*Options, error) {
 }
 
 func NewCmdSnapshot(log logrus.FieldLogger) *cli.Command { //nolint:funlen
-	var o *Options
 	return &cli.Command{
 		Name:        "snapshot",
 		Usage:       "Manage snapshots of your developer environment",
 		Description: cmdutil.NewDescription(snapshotLongDesc, helpersExample),
-		Before: func(c *cli.Context) error {
-			var err error
-			o, err = NewOptions(log)
-			return err
-		},
 		Subcommands: []*cli.Command{
 			{
 				Name:        "generate",
@@ -118,13 +127,23 @@ func NewCmdSnapshot(log logrus.FieldLogger) *cli.Command { //nolint:funlen
 					},
 				},
 				Action: func(c *cli.Context) error {
-					b, err := ioutil.ReadFile("snapshots.yaml")
+					b, err := box.LoadBox()
+					if err != nil {
+						return errors.Wrap(err, "failed to load box configuration")
+					}
+
+					o, err := NewOptions(log, b)
+					if err != nil {
+						return err
+					}
+
+					byt, err := ioutil.ReadFile("snapshots.yaml")
 					if err != nil {
 						return err
 					}
 
 					var s *box.SnapshotGenerateConfig
-					err = yaml.Unmarshal(b, &s)
+					err = yaml.Unmarshal(byt, &s)
 					if err != nil {
 						return err
 					}
@@ -136,269 +155,333 @@ func NewCmdSnapshot(log logrus.FieldLogger) *cli.Command { //nolint:funlen
 	}
 }
 
-func (o *Options) DeleteSnapshot(ctx context.Context, snapshotName string) error {
-	if o.vc == nil {
-		return fmt.Errorf("velero client not set")
+// awsEndpointResolver is a stub aws.EndpointResolver that returns a static
+// endpoint.
+type awsEndpointResolver struct {
+	endpoint string
+	region   string
+}
+
+func (a *awsEndpointResolver) ResolveEndpoint(_, _ string) (aws.Endpoint, error) {
+	return aws.Endpoint{
+		PartitionID:       "aws",
+		URL:               a.endpoint,
+		HostnameImmutable: true,
+		SigningRegion:     a.region,
+	}, nil
+}
+
+func (o *Options) Generate(ctx context.Context, s *box.SnapshotGenerateConfig,
+	skipUpload bool, channel box.SnapshotLockChannel) error { //nolint:funlen
+	o.log.WithField("snapshots", len(s.Targets)).Info("Generating Snapshots")
+
+	copts := devenvaws.DefaultCredentialOptions()
+	if o.b.DeveloperEnvironmentConfig.SnapshotConfig.WriteAWSRole != "" {
+		copts.Role = o.b.DeveloperEnvironmentConfig.SnapshotConfig.WriteAWSRole
+	}
+	copts.Log = o.log
+	err := devenvaws.EnsureValidCredentials(ctx, copts)
+	if err != nil {
+		return errors.Wrap(err, "failed to get necessary permissions")
 	}
 
-	if snapshotName == "" {
-		return fmt.Errorf("missing snapshot name")
-	}
-
-	_, err := o.GetSnapshot(ctx, snapshotName)
+	cfg, err := config.LoadDefaultConfig(ctx, config.WithRegion(o.b.DeveloperEnvironmentConfig.SnapshotConfig.Region))
 	if err != nil {
 		return err
 	}
 
-	_, err = o.vc.VeleroV1().DeleteBackupRequests(SnapshotNamespace).Create(ctx, &velerov1api.DeleteBackupRequest{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: snapshotName + "-",
-		},
-		Spec: velerov1api.DeleteBackupRequestSpec{
-			BackupName: snapshotName,
-		},
-	}, metav1.CreateOptions{})
+	// use a custom endpoint if provided
+	if endpoint := o.b.DeveloperEnvironmentConfig.SnapshotConfig.Endpoint; endpoint != "" {
+		cfg.EndpointResolver = &awsEndpointResolver{ //nolint:staticcheck // Why: using new one doesn't work?
+			endpoint: endpoint,
+			region:   cfg.Region,
+		}
+	}
 
+	o.log.WithField("region", cfg.Region).Info("s3 config")
+
+	s3c := s3.NewFromConfig(cfg)
+
+	lockfile := &box.SnapshotLock{}
+	resp, err := s3c.GetObject(ctx, &s3.GetObjectInput{
+		Bucket: &o.b.DeveloperEnvironmentConfig.SnapshotConfig.Bucket,
+		Key:    aws.String("automated-snapshots/v2/latest.yaml"),
+	})
+	if err == nil {
+		defer resp.Body.Close()
+		err = yaml.NewDecoder(resp.Body).Decode(&lockfile)
+		if err != nil {
+			return errors.Wrap(err, "failed to parse remote snapshot lockfile")
+		}
+	} else {
+		o.log.WithError(err).
+			Warn("Failed to fetch existing remote snapshot lockfile, will generate a new one")
+	}
+
+	if lockfile.TargetsV2 == nil {
+		lockfile.TargetsV2 = make(map[string]*box.SnapshotLockList)
+	}
+
+	for name, t := range s.Targets {
+		//nolint:govet // Why: We're OK shadowing err
+		itm, err := o.generateSnapshot(ctx, s3c, name, t, skipUpload)
+		if err != nil {
+			return err
+		}
+
+		if _, ok := lockfile.TargetsV2[name]; !ok {
+			lockfile.TargetsV2[name] = &box.SnapshotLockList{}
+		}
+
+		if lockfile.TargetsV2[name].Snapshots == nil {
+			lockfile.TargetsV2[name].Snapshots = make(map[box.SnapshotLockChannel][]*box.SnapshotLockListItem)
+		}
+
+		if _, ok := lockfile.TargetsV2[name].Snapshots[channel]; !ok {
+			lockfile.TargetsV2[name].Snapshots[channel] = make([]*box.SnapshotLockListItem, 0)
+		}
+
+		// Make this the latest version
+		lockfile.TargetsV2[name].Snapshots[channel] = append(
+			[]*box.SnapshotLockListItem{itm}, lockfile.TargetsV2[name].Snapshots[channel]...,
+		)
+	}
+
+	// Don't generate a lock if we're not uploading
+	if skipUpload {
+		return nil
+	}
+
+	lockfile.GeneratedAt = time.Now().UTC()
+
+	byt, err := yaml.Marshal(lockfile)
+	if err != nil {
+		return err
+	}
+
+	_, err = s3c.PutObject(ctx, &s3.PutObjectInput{
+		Bucket: &o.b.DeveloperEnvironmentConfig.SnapshotConfig.Bucket,
+		Key:    aws.String(filepath.Join("automated-snapshots", "v2", "latest.yaml")),
+		Body:   bytes.NewReader(byt),
+	})
 	return err
 }
 
-func (o *Options) GetSnapshot(ctx context.Context, snapshotName string) (*velerov1api.Backup, error) {
-	if o.vc == nil {
-		return nil, fmt.Errorf("velero client not set")
+//nolint:funlen,gocritic
+func (o *Options) uploadSnapshot(ctx context.Context, s3c *s3.Client,
+	name string, t *box.SnapshotTarget) (string, string, error) {
+	tmpFile, err := os.CreateTemp("", "snapshot-*")
+	if err != nil {
+		return "", "", err
+	}
+	defer os.Remove(tmpFile.Name())
+
+	hash := md5.New() //nolint:gosec // Why: We're just creating a digest
+	tw := tar.NewWriter(io.MultiWriter(tmpFile, hash))
+
+	o.k, err = kube.GetKubeClient()
+	if err != nil {
+		return "", "", err
 	}
 
-	return o.vc.VeleroV1().Backups(SnapshotNamespace).Get(ctx, snapshotName, metav1.GetOptions{})
-}
+	mc, err := snapshoter.NewSnapshotBackend(ctx, o.r, o.k)
+	if err != nil {
+		return "", "", err
+	}
 
-func (o *Options) deleteNamespaces(ctx context.Context) error { //nolint:funlen
-	var namespaces []interface{}
-	cont := ""
-	for {
-		l, err := o.k.CoreV1().Namespaces().List(ctx, metav1.ListOptions{
-			Continue: cont,
+	o.log.Info("creating tar archive")
+	for obj := range mc.ListObjects(ctx, noncmdsnapshot.SnapshotNamespace, minio.ListObjectsOptions{Recursive: true}) {
+		// Skip empty keys
+		if strings.EqualFold(obj.Key, "") {
+			continue
+		}
+
+		sObj, err := mc.GetObject(ctx, noncmdsnapshot.SnapshotNamespace, obj.Key, minio.GetObjectOptions{}) //nolint:govet
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to get object from local S3")
+		}
+
+		info, err := sObj.Stat()
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to stat object")
+		}
+
+		err = tw.WriteHeader(&tar.Header{
+			Typeflag:   tar.TypeReg,
+			Name:       info.Key,
+			Size:       info.Size,
+			Mode:       0755,
+			ModTime:    info.LastModified,
+			AccessTime: info.LastModified,
+			ChangeTime: info.LastModified,
 		})
 		if err != nil {
-			return err
-		}
-		cont = l.Continue
-
-		for i := range l.Items {
-			namespaces = append(namespaces, &l.Items[i])
+			return "", "", errors.Wrap(err, "failed to write tar header")
 		}
 
-		if l.Continue == "" {
-			break
-		}
-	}
-
-	if _, err := worker.ProcessArray(ctx, namespaces, func(ctx context.Context, itm interface{}) (interface{}, error) {
-		n := itm.(*corev1.Namespace)
-
-		// skip some namespaces
-		switch n.Name {
-		case "default", "kube-system", "velero", "kube-public", "kube-node-lease", "nginx-ingress", "local-path-storage":
-			return nil, nil
-		}
-
-		log := o.log.WithField("namespace", n.Name)
-		log.Info("deleting namespace")
-		err := o.k.CoreV1().Namespaces().Delete(ctx, n.Name, metav1.DeleteOptions{})
+		_, err = io.Copy(tw, sObj)
 		if err != nil {
-			return nil, err
+			return "", "", errors.Wrap(err, "failed to download object from local S3")
 		}
-
-		ticker := time.NewTicker(5 * time.Second)
-	loop:
-		for {
-			select {
-			case <-ctx.Done():
-				return nil, ctx.Err()
-			case <-ticker.C:
-				_, err = o.k.CoreV1().Namespaces().Get(ctx, n.Name, metav1.GetOptions{})
-				if kerrors.IsNotFound(err) {
-					break loop
-				} else if err != nil {
-					return nil, err
-				}
-
-				log.Info("waiting for namespace to delete ...")
-			}
-		}
-
-		log.Info("deleted namespace")
-		return nil, nil
-	}); err != nil {
-		return err
 	}
 
-	return nil
+	// If we have post-restore manifests, then include them in the archive at a well-known
+	// path for post-processing on runtime
+	if t.PostRestore != "" {
+		f, err := os.Open(t.PostRestore) //nolint:govet // Why: We're OK shadowing err.
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to open post-restore file")
+		}
+
+		inf, err := f.Stat()
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to stat post-restore file")
+		}
+
+		header, err := tar.FileInfoHeader(inf, "")
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to create tar header")
+		}
+		header.Name = "post-restore/manifests.yaml"
+
+		err = tw.WriteHeader(header)
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to write tar header")
+		}
+
+		_, err = io.Copy(tw, f)
+		if err != nil {
+			return "", "", errors.Wrap(err, "failed to write post-restore file to archive")
+		}
+	}
+
+	if err := tw.Close(); err != nil { //nolint:govet // Why: we're OK shadowing err
+		return "", "", err
+	}
+	if err := tmpFile.Close(); err != nil { //nolint:govet // Why: we're OK shadowing err
+		return "", "", err
+	}
+
+	hashStr := base64.StdEncoding.EncodeToString(hash.Sum(nil))
+	key := filepath.Join("automated-snapshots", "v2", name, strconv.Itoa(int(time.Now().UTC().UnixNano()))+".tar")
+
+	tmpFile, err = os.Open(tmpFile.Name())
+	if err != nil {
+		return "", "", err
+	}
+
+	obj := &s3.PutObjectInput{
+		Bucket:     &o.b.DeveloperEnvironmentConfig.SnapshotConfig.Bucket,
+		Key:        &key,
+		Body:       tmpFile,
+		ContentMD5: &hashStr,
+	}
+
+	o.log.WithField("bucket", *obj.Bucket).WithField("key", *obj.Key).Info("uploading tar archive")
+	_, err = s3c.PutObject(ctx, obj)
+	if err != nil {
+		return "", "", err
+	}
+
+	return hashStr, key, nil
 }
 
-func (o *Options) deleteExistingRestore(ctx context.Context, snapshotName string) error {
-	restore, err := o.vc.VeleroV1().Restores(SnapshotNamespace).Get(ctx, snapshotName, metav1.GetOptions{})
-	if err == nil {
-		if restore.Status.Phase == velerov1api.RestorePhaseInProgress {
-			return fmt.Errorf("existing restore is in progress, refusing to create new restore")
+//nolint:funlen
+func (o *Options) generateSnapshot(ctx context.Context, s3c *s3.Client,
+	name string, t *box.SnapshotTarget, skipUpload bool) (*box.SnapshotLockListItem, error) {
+	o.log.WithField("snapshot", name).Info("Generating Snapshot")
+
+	destroyOpts, err := destroy.NewOptions(o.log)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create destroy command")
+	}
+	destroyOpts.Run(ctx) //nolint:errcheck
+
+	// using exec because of an import cycle, need to fix
+	os.Setenv("DEVENV_SNAPSHOT_GENERATION", "true") //nolint:errcheck
+
+	popts, err := provision.NewOptions(o.log, o.b)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create options for provision")
+	}
+
+	popts.KubernetesRuntime, err = kubernetesruntime.GetRuntime("kind")
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create kind runtime")
+	}
+	popts.Base = true
+
+	if err := popts.Run(ctx); err != nil {
+		return nil, errors.Wrap(err, "failed to provision devenv")
+	}
+
+	if len(t.DeployApps) != 0 {
+		o.log.Info("Deploying applications into devenv")
+		for _, app := range t.DeployApps {
+			o.log.WithField("application", app).Info("Deploying application")
+			cmd := exec.CommandContext(ctx, os.Args[0], "--skip-update", "deploy-app", app) //nolint:gosec
+			cmd.Stderr = os.Stderr
+			cmd.Stdout = os.Stdout
+			cmd.Stdin = os.Stdin
+			if err := cmd.Run(); err != nil { //nolint:govet // Why: We're OK shadowing err.
+				return nil, errors.Wrap(err, "failed to deploy application")
+			}
 		}
-		o.log.Info("Deleting previous completed restore")
-		err = o.vc.VeleroV1().Restores(SnapshotNamespace).Delete(ctx, snapshotName, metav1.DeleteOptions{})
+	}
+
+	if t.Command != "" {
+		o.log.Info("Running snapshot generation command")
+		err = cmdutil.RunKubernetesCommand(ctx, "", false, "/bin/bash", "-c", t.Command)
 		if err != nil {
-			return errors.Wrap(err, "failed to delete existing restore")
+			return nil, errors.Wrap(err, "failed to run snapshot supplied command")
 		}
+	}
 
-		o.log.Info("Waiting for delete to finish ...")
-		ticker := time.NewTicker(5 * time.Second)
-	loop:
-		for {
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case <-ticker.C:
-				_, err = o.vc.VeleroV1().Restores(SnapshotNamespace).Get(ctx, snapshotName, metav1.GetOptions{})
-				if kerrors.IsNotFound(err) {
-					break loop
-				} else if err != nil {
-					return err
-				}
+	if len(t.PostDeployApps) != 0 {
+		o.log.Info("Deploying applications into devenv")
+		for _, app := range t.PostDeployApps {
+			o.log.WithField("application", app).Info("Deploying application")
+			cmd := exec.CommandContext(ctx, os.Args[0], "--skip-update", "deploy-app", app) //nolint:gosec
+			cmd.Stderr = os.Stderr
+			cmd.Stdout = os.Stdout
+			cmd.Stdin = os.Stdin
+			if err := cmd.Run(); err != nil { //nolint:govet // Why: We're OK shadowing err.
+				return nil, errors.Wrap(err, "failed to deploy application")
 			}
 		}
 	}
 
-	return nil
-}
-
-func (o *Options) RestoreSnapshot(ctx context.Context, snapshotName string, liveRestore bool) error { //nolint:funlen
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	if snapshotName == "" {
-		return fmt.Errorf("missing snapshot name")
+	// Need to create a new Kubernetes client that uses the new cluster
+	o, err = NewOptions(o.log, o.b)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create new clients")
 	}
 
-	if liveRestore {
-		o.log.Warn("THIS WILL DELETE ALL EXISTING DATA IN YOUR CLUSTER FROM BEFORE THE SNAPSHOT. PROCEED?")
-		proceed, err := cmdutil.GetYesOrNoInput(ctx)
+	err = devenvutil.WaitForAllPodsToBeReady(ctx, o.k, o.log)
+	if err != nil {
+		return nil, err
+	}
+
+	veleroBackupName, err := o.CreateSnapshot(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	hash := "unknown"
+	key := "unknown"
+	if !skipUpload {
+		hash, key, err = o.uploadSnapshot(ctx, s3c, name, t)
 		if err != nil {
-			return err
-		}
-
-		if !proceed {
-			return fmt.Errorf("denied to proceed")
+			return nil, errors.Wrap(err, "failed to upload snapshot")
 		}
 	}
 
-	if _, err := o.GetSnapshot(ctx, snapshotName); err != nil {
-		return err
-	}
-
-	if err := o.deleteExistingRestore(ctx, snapshotName); err != nil {
-		return err
-	}
-
-	if liveRestore {
-		if err := o.deleteNamespaces(ctx); err != nil {
-			return err
-		}
-	}
-
-	if _, err := o.vc.VeleroV1().Restores(SnapshotNamespace).Create(ctx, &velerov1api.Restore{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: SnapshotNamespace,
-			Name:      snapshotName,
-		},
-		Spec: velerov1api.RestoreSpec{
-			BackupName:              snapshotName,
-			RestorePVs:              boolptr.True(),
-			IncludeClusterResources: boolptr.True(),
-			PreserveNodePorts:       boolptr.True(),
-
-			// TODO(DTSS-829): This should be moved into the generation framework
-			ExcludedNamespaces: []string{
-				"nginx-ingress",
-				"kube-system",
-				"cert-manager",
-				"nginx-ingress",
-				"velero",
-				"minio",
-				"vault-secrets-operator",
-				"local-path-storage",
-				"monitoring",
-				"resourcer--bento1a",
-			},
-		},
-	}, metav1.CreateOptions{}); err != nil {
-		return err
-	}
-
-	updates := make(chan *velerov1api.Restore)
-	restoreInformer := velerov1.NewRestoreInformer(o.vc, SnapshotNamespace, 0, nil)
-	restoreInformer.AddEventHandler( //nolint:dupl
-		cache.FilteringResourceEventHandler{
-			FilterFunc: func(obj interface{}) bool {
-				restore, ok := obj.(*velerov1api.Restore)
-				if !ok {
-					return false
-				}
-				return restore.Name == snapshotName
-			},
-			Handler: cache.ResourceEventHandlerFuncs{
-				UpdateFunc: func(_, obj interface{}) {
-					restore, ok := obj.(*velerov1api.Restore)
-					if !ok {
-						return
-					}
-					updates <- restore
-				},
-				DeleteFunc: func(obj interface{}) {
-					restore, ok := obj.(*velerov1api.Restore)
-					if !ok {
-						return
-					}
-					updates <- restore
-				},
-			},
-		},
-	)
-	go restoreInformer.Run(ctx.Done())
-
-	o.log.Info("Waiting for snapshot restore operation to complete ...")
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case restore, ok := <-updates:
-			if !ok {
-				return fmt.Errorf("failed to watch restore operation")
-			}
-			if restore.Status.Phase != velerov1api.RestorePhaseNew && restore.Status.Phase != velerov1api.RestorePhaseInProgress {
-				o.log.Infof("Snapshot restore finished with status: %v", restore.Status.Phase)
-				return nil
-			}
-		}
-	}
-}
-
-// CreateBackupStorage creates a backup storage location
-func (o *Options) CreateBackupStorage(ctx context.Context, name, bucket string) error {
-	_, err := o.vc.VeleroV1().BackupStorageLocations(SnapshotNamespace).Create(ctx, &velerov1api.BackupStorageLocation{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: velerov1api.BackupStorageLocationSpec{
-			Provider: "aws",
-			StorageType: velerov1api.StorageType{
-				ObjectStorage: &velerov1api.ObjectStorageLocation{
-					Bucket: bucket,
-				},
-			},
-			Config: map[string]string{
-				"region":           "minio",
-				"s3ForcePathStyle": "true",
-				"s3Url":            "http://minio.minio:9000",
-			},
-		},
-	}, metav1.CreateOptions{})
-	return err
+	return &box.SnapshotLockListItem{
+		Digest:           hash,
+		URI:              key,
+		Config:           t,
+		VeleroBackupName: veleroBackupName,
+	}, nil
 }
 
 func (o *Options) CreateSnapshot(ctx context.Context) (string, error) { //nolint:funlen
@@ -406,7 +489,7 @@ func (o *Options) CreateSnapshot(ctx context.Context) (string, error) { //nolint
 	defer cancel()
 
 	updates := make(chan *velerov1api.Backup)
-	backupInformer := velerov1.NewBackupInformer(o.vc, SnapshotNamespace, 0, nil)
+	backupInformer := velerov1.NewBackupInformer(o.vc, noncmdsnapshot.SnapshotNamespace, 0, nil)
 
 	// Create DNS1133 compliant backup name.
 	backupName := strings.ToLower(
@@ -442,7 +525,7 @@ func (o *Options) CreateSnapshot(ctx context.Context) (string, error) { //nolint
 	)
 	go backupInformer.Run(ctx.Done())
 
-	_, err := o.vc.VeleroV1().Backups(SnapshotNamespace).Create(ctx, &velerov1api.Backup{
+	_, err := o.vc.VeleroV1().Backups(noncmdsnapshot.SnapshotNamespace).Create(ctx, &velerov1api.Backup{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: backupName,
 		},
